+++ ValueFnIter/FHorz/ExperienceAssetze/DivideConquerGridInterpLayer/ValueFnIter_FHorz_ExpAssetze_DC1_GI1_nod1_e_raw.m
@@ -1,6 +1,6 @@
 function [V,Policy]=ValueFnIter_FHorz_ExpAssetze_DC1_GI1_nod1_e_raw(n_d2,n_a1,n_a2,n_z,n_e,N_j, d2_gridvals, a1_gridvals, a2_grid, z_gridvals_J, e_gridvals_J, pi_z_J, pi_e_J, ReturnFn, aprimeFn, Parameters, DiscountFactorParamNames, ReturnFnParamNames, aprimeFnParamNames, vfoptions)
 
-N_d2=double(prod(n_d2));
+N_d2=prod(n_d2);
 N_a1=prod(n_a1);
 N_a2=prod(n_a2);
 N_a=N_a1*N_a2;
